fix(recvpn): make u64-to-int bit extraction explicit
below_seen returns a single extracted bit as int; add an explicit cast
to signal the intended narrowing (bugprone-narrowing-conversions).

+++ src/recvpn/recvpn.c
@@ -12,7 +12,7 @@ static int below_seen(const quic_recvpn *r, u64 pn)
 {
     u64 delta = r->largest - pn;
     if (delta > QUIC_RECVPN_WINDOW) return 0; /* fell out of the window */
-    return (r->bitmap >> (delta - 1)) & 1u;
+    return (int)((r->bitmap >> (delta - 1)) & 1u);
 }
 
 /* Classify pn against largest: 1 seen (==), 0 not (>), else check the window. */
